+++ R/simulation-simInit.R
@@ -357,8 +357,19 @@ setMethod(
 
     # user modules
     modulesLoaded <- list()
+
+    # If this is being run inside a module, then it needs to know
+    simPrev <- .grepSysCalls(sys.calls(), ".runEvent")
+
     # create simList object for the simulation
     sim <- new("simList")
+
+    if (length(simPrev) > 0) {
+      sim$._simPrevs <- append(sys.frames()[tail(simPrev, 1)], sim$._simPrev)
+    } else {
+      sim$._simPrevs <- list()
+    }
+
     # Make a temporary place to store parsed module files
     sim@.xData[[".parsedFiles"]] <- new.env(parent = emptyenv())
     on.exit(rm(".parsedFiles", envir = sim@.xData), add = TRUE )
@@ -1096,7 +1107,7 @@ simInitAndSpades <- function(times, params, modules, objects, paths, inputs, out
   allObjsProvided <- sim@depends@dependencies[[i]]@inputObjects[["objectName"]] %in%
     sim$.userSuppliedObjNames
   if (!all(allObjsProvided)) {
-    if (!is.null(sim@.xData$.mods[[mBase]][[".inputObjects"]])) {
+    if (!is.null(.getModuleInputObjects(sim, m))) {
       # browser(expr = exists("._runModuleInputObjects_2"))
       if (!missing(objects))
         list2env(objects[sim@depends@dependencies[[i]]@inputObjects[["objectName"]][allObjsProvided]],