+++ tests/testthat/test-mod.R
@@ -7,9 +7,11 @@ test_that("local mod object", {
 
   newModule("test", tmpdir, open = FALSE)
   newModule("test2", tmpdir, open = FALSE)
+  testFilePath <- file.path(tmpdir, "test", "test.R")
+  test2FilePath <- file.path(tmpdir, "test2", "test2.R")
 
   # Sept 18 2018 -- Changed to use "seconds" -- better comparison with simple loop
-  cat(file = file.path(tmpdir, "test", "test.R"),'
+  cat(file = testFilePath,'
       defineModule(sim, list(
       name = "test",
       description = "insert module description here",
@@ -41,6 +43,10 @@ test_that("local mod object", {
       mod$a <- 2
       sim$testPar1 <- Par$testParA
 
+      if (tryCatch(exists("Init", envir = asNamespace("test"), inherits = FALSE), error = function(x) FALSE)) {
+        sim <- Init(sim)
+      }
+
       sim <- scheduleEvent(sim, sim@simtimes[["current"]] + 1, "test", "event1", .skipChecks = TRUE)
       },
       event1 = {
@@ -56,7 +62,7 @@ test_that("local mod object", {
       }
       ', fill = TRUE)
 
-  cat(file = file.path(tmpdir, "test2", "test2.R"),'
+  cat(file = test2FilePath,'
       defineModule(sim, list(
       name = "test2",
       description = "insert module description here",
@@ -91,11 +97,24 @@ test_that("local mod object", {
       switch(
       eventType,
       init = {
+      if (tryCatch(exists("Init", envir = asNamespace("test2"), inherits = FALSE), error = function(x) FALSE)) {
+        sim <- Init(sim)
+      }
+
       if (isTRUE(P(sim)$testParB >= 1100)) {
          P(sim, "testParB") <-  P(sim)$testParB + 756
       }
 
-
+      if (any(grepl("testCommonPar", names(unlist(params(sim)))))) {
+            errorText <- try(paramCheckOtherMods(sim, "testCommonPar"), silent = TRUE)
+            if (identical("try-error", attr(errorText, "class")))
+              message("There was an error")
+            warn <- capture_warnings(paramCheckOtherMods(sim, "testCommonPar", ifSetButDifferent = "warning"))
+            if (length(warn))
+              message("There was a warning")
+            paramCheckOtherMods(sim, "testCommonPar", ifSetButDifferent = "message")
+            paramCheckOtherMods(sim, "testCommonPar", ifSetButDifferent = "silent")
+      }
       if (isTRUE(!is.null(P(sim)$testRestartSpades))) {
         stop("testing restartSpades")#browser()
       }
@@ -216,19 +235,124 @@ test_that("local mod object", {
   expect_true(identical(P(mySim7, module = "test2", "testParF"), 77))
   expect_true(identical(P(mySim7, module = "test2", "testParG"), 79))
   expect_true(identical(P(mySim7, module = "test2", "testParH"), 48))
-  # expect_true(identical(P(mySim7, module = "test2", "testParAB"), 234))
+
+  # Test common parameters i.e., globals
+  # Set one to NULL
+  vals <- list("sdfd", "default", NULL, "ffff")
+  lens <- list(0,0,0,1)
+  out <- Map(len = lens, val = vals, function(len, val) {
+    mySim11 <- simInit(times = list(start = 0, end = 0),
+                       paths = list(modulePath = tmpdir), modules = c("test", "test2"),
+                       params = list(test = list(testCommonPar = vals[[1]]),
+                                     test2 = list(testCommonPar = val)))
+    mess <- capture_messages(spades(mySim11, debug = FALSE))
+    expect_true(sum(grepl("multiple values", mess)) == len)
+    expect_true(sum(grepl("There was a warning", mess)) == len)
+    expect_true(sum(grepl("There was an error", mess)) == len)
+  })
+
+
+
 
   # Test restartSpades # The removal of the completed ... it shouldn't, but it did previously
   if (interactive()) {
+    opt <- options("spades.recoveryMode" = TRUE)
+    on.exit(opt, add = TRUE)
     mySim8 <- simInit(times = list(start = 0, end = 0),
                       paths = list(modulePath = tmpdir), modules = c("test", "test2"),
                       params = list(test2 = list(testRestartSpades = 1)))
-    err <- capture_error(ss <- spades(mySim8))
+    err <- try(ss <- spades(mySim8, debug = FALSE), silent = TRUE)
 
-    err <- capture_error(sim2 <- restartSpades(.pkgEnv$.sim)) # is missing completed events
-    err <- capture_error(sim3 <- restartSpades(.pkgEnv$.sim)) # is missing completed events
-    .pkgEnv$.sim@params$test2$testRestartSpades <- NULL
-    sim3 <- restartSpades(.pkgEnv$.sim)
+    sim <- asNamespace("SpaDES.core")$.pkgEnv$.sim
+    err <- capture_error(sim2 <- restartSpades(sim, debug = FALSE)) # is missing completed events
+
+    sim <- asNamespace("SpaDES.core")$.pkgEnv$.sim
+    err <- capture_error(sim3 <- restartSpades(sim, debug = FALSE)) # is missing completed events
+
+    sim <- asNamespace("SpaDES.core")$.pkgEnv$.sim
+    sim@params$test2$testRestartSpades <- NULL
+    sim3 <- restartSpades(sim, debug = FALSE)
     expect_true(NROW(completed(sim3)) == 7)
+    options("spades.recoveryMode" = FALSE)
   }
+
+
+  # Test converting these to packages
+  if (interactive()) {
+    if (requireNamespace("pkgload")) {
+      cat(file = testFilePath,'
+      Init <- function(sim) {
+        sim$aaaa <- Run(1)
+        return(sim)
+      }
+
+      Run <- function(a) {
+        return(a + 1)
+      }
+      ', fill = TRUE, append = TRUE)
+
+      cat(file = test2FilePath,'
+      Init <- function(sim) {
+        sim$cccc <- try(Run(1), silent = TRUE)
+        return(sim)
+      }
+
+      Run2 <- function(a) {
+        return(a + 2)
+      }
+      ', fill = TRUE, append = TRUE)
+
+      # aaaaa <<- 1
+
+      for (tt in c("test", "test2")) {
+        expect_true(!file.exists(file.path(tmpdir, tt, "DESCRIPTION")))
+        expect_true(!file.exists(file.path(tmpdir, tt, "NAMESPACE")))
+      }
+      convertToPackage(module = "test", path = tmpdir)
+      convertToPackage(module = "test2", path = tmpdir)
+      for (tt in c("test", "test2")) {
+        expect_true(file.exists(file.path(tmpdir, tt, "DESCRIPTION")))
+        expect_true(!file.exists(file.path(tmpdir, tt, "NAMESPACE")))
+        expect_true(dir.exists(file.path(tmpdir, tt, "R")))
+      }
+      expect_true(file.exists(file.path(tmpdir, "test", "R", "Init.R")))
+      expect_true(file.exists(file.path(tmpdir, "test2", "R", "Init.R")))
+      expect_true(file.exists(file.path(tmpdir, "test", "R", "Run.R")))
+      expect_true(file.exists(file.path(tmpdir, "test2", "R", "Run2.R")))
+
+      mySim9 <- simInit(times = list(start = 0, end = 0),
+                        paths = list(modulePath = tmpdir), modules = c("test", "test2"))
+
+      # doesn't document
+      for (tt in c("test", "test2")) {
+        expect_true(file.exists(file.path(tmpdir, tt, "DESCRIPTION")))
+        expect_true(!file.exists(file.path(tmpdir, tt, "NAMESPACE")))
+        expect_true(dir.exists(file.path(tmpdir, tt, "R")))
+      }
+      working <- spades(mySim9, debug = FALSE)
+
+      # document -- this exports all functions!! Danger for testing later
+      out <- lapply(c("test", "test2"), function(tt) {
+        roxygen2::roxygenise(file.path(tmpdir, tt))
+      })
+
+      # Will run document() so will have the NAMESPACE and
+      for (tt in c("test", "test2")) {
+        expect_true(file.exists(file.path(tmpdir, tt, "DESCRIPTION")))
+        expect_true(file.exists(file.path(tmpdir, tt, "NAMESPACE")))
+        expect_true(sum(grepl("export.+doEvent", readLines(file.path(tmpdir, tt, "NAMESPACE")))) == 1)
+      }
+
+      # check that inheritance is correct -- Run is in the namespace, Init also... doEvent calls Init calls Run
+      expect_true(is(working, "simList"))
+      expect_true(working$aaaa == 2)
+      expect_true(is(working$cccc, "try-error"))
+      bbb <- Run2(2)
+      expect_true(bbb == 4)
+      pkgload::unload("test")
+      pkgload::unload("test2")
+    }
+  }
+
+
 })