+++ NEWS.md
@@ -1,5 +1,31 @@
 Known issues: <https://github.com/PredictiveEcology/SpaDES.core/issues>
 
+version 1.0.10
+==============
+
+## new features
+* **experimental new feature** `SpaDES` modules can now be R packages.
+  The simplest way to convert a module to a package is using the new function `convertToPackage`.
+  Benefits of doing this are so that a SpaDES module can benefit from the infrastructure of an R package (e.g., `devtools::document`, `devtools::check`, setting up Continuous Integration systems etc.).
+  Any documentation written using `#'` i.e., `roxygen2` will be copied immediately above the function where it was sitting. 
+* `newModule` now correctly places the `SpaDES.core` package dependency in the `reqdPkgs` element of the metadata, instead of `version`. It will put the full Github reference if SpaDES.core was installed directly from GitHub.
+* There is a bug in `qs` package: either `qsave` or `qread` converts `data.table` objects to `list` objects. `loadSimList` has a work around internally to convert these objects back to `data.table`, if the metadata indicate that the objects should be `data.table` objects.
+* new function: `paramCheckOtherMods`.
+  Can be used within a module to assert that a parameter has the same value as the same parameter in other modules.
+  This is therefore a check of a parameter that might be considered a `.global` and passed within `simInit(..., params = list(.globals = list(someParam = "someValue")))`, but the user did not do that.
+* now `spades` messaging when e.g., `debug = 1` can correctly accommodate nested `spades` calls, i.e., a SpaDES module calling `spades` internally.
+* `newModule` now puts `SpaDES.core` dependency in the correct `reqdPkgs` instead of `version` metadata element
+* to further the transition to using `.plots` instead of `.plotInitialTime`, `Plots` will check whether `.plotInitialTime` is actually set in the module metadata first.
+  Only if it is there, will it evaluate its value. Currently, modules get default values for `.plotInitialTime` even if the module developer didn't include it in the module metadata. 
+
+## dependency changes
+* drop support for R 3.6 (#178)
+
+## bug fixes
+* minor bugfix when `debug` arg of `spades` is set to an event type that is also in the core modules (e.g., save, load), such as "init"
+* `Cache`-ing of a `simList`, when `quick` is a character vector, errored. Now fixed.
+* `moduleMetadata` incorrectly dropped `defineModuleListItems` under certain signatures.
+
 version 1.0.9
 =============